requirements.txt
dataclasses-json
pillow; platform_machine != "x86_64"
pillow-simd; platform_machine == "x86_64"